+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 *EXOSharedMailbox
   * Fixed an issue where Aliases were not properly removed
     (Issue #749);
+* TeamsCallingPolicy
+  * Added support for the AllowWebPSTNCalling property;
 * TeamsChannel
   * Fixed an error in the Export when trying to connect using
     only an Azure AD Application without any credentials